Update tree_node.json
We propose an icon for the tag leaf_type:palm , it is more representative for palm species and others of similar shape. The leaf_type:broadleaved tag is not the best icon for these cases. As a background, the Osm app has three types of icons, it recognises leaf_type:palm. Regards